requirements.txt
geopy==2.3.0
openpyxl==3.1.2
pandas==2.0.3
pillow==9.5.0
pip-chill==1.0.3
psycopg2-binary==2.9.6
pylogrus==0.4.0
python-dotenv==1.0.0
telebot==0.0.5
xlsxwriter==3.1.2
